Attempt to implement CountDownLatch. The intention was to use CountDownLatch instead of waiting for tests. For instance, when a task is given to an executor, a thread is not spawned instantly. Then, we wait for 1 millisecond for the thread to be spawned and them make an assertion about the number of waiting threads. Maybe there is a better way to test the functions that required waiting.
To prevent the usage of Thread.sleep(), I was suggested to use CountDownLatch instead.
However, CountDownLatch doesn't solve the stated problem of threads spawning with a slight delay.
In addition, it seems that I will need more than one CountDownLatch in the main function to test it.

This makes the main logic even more obscure than it already is with the waiting delay and timeout.

There fore, the attempt to implement CountDownLatch was submitted as a demonstration of what exactly was attempted.

This version fails the concurrent tests when the latch is checked right after submitting tasks to the executor.